requirements.txt
amqp==5.0.6
asgiref==3.4.1
backcall==0.2.0
billiard==3.6.4.0
black==21.11b1
cached-property==1.5.2
celery==5.2.1
certifi==2021.10.8
charset-normalizer==2.0.7
click==8.0.3
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.2.0
decorator==5.1.0
Deprecated==1.2.13
Django==3.2.9
django-extensions==3.1.5
django-filter==21.1
django-rest-auth==0.9.5
django-rest-passwordreset==1.2.1
djangorestframework==3.12.4
djangorestframework-jwt==1.11.0
# djangorestframework-simplejwt==5.0.0
idna==3.3
importlib-metadata==4.8.2
ipython==7.30.1
jedi==0.18.1
kombu==5.2.2
matplotlib-inline==0.1.3
mypy-extensions==0.4.3
parso==0.8.3
pathspec==0.9.0
pexpect==4.8.0
pickleshare==0.7.5
platformdirs==2.4.0
prompt-toolkit==3.0.23
psycopg2-binary==2.9.2
ptyprocess==0.7.0
Pygments==2.10.0
# PyJWT==1.7.1
# pytz==2021.3
razorpay==1.2.0
redis==4.0.2
regex==2021.11.10
requests==2.26.0
six==1.16.0
sqlparse==0.4.2
tomli==1.2.2
traitlets==5.1.1
typed-ast==1.5.0
typing_extensions==4.0.0
urllib3==1.26.7
vine==5.0.0
wcwidth==0.2.5
wrapt==1.13.3
zipp==3.6.0
